requirements.txt
alembic==1.7.5
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2==2.9.3
python-dotenv==0.19.2
requests==2.26.0
SQLAlchemy==1.4.28
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==3.0.0
zipp==3.6.0
